moved update user info form into a separate component
